% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modele_additif.R
\name{modele_additif}
\alias{modele_additif}
\title{Décomposition additive}
\usage{
modele_additif(x)
}
\arguments{
\item{x}{objet ts (fréquence >= 2)}
}
\value{
objet decompose
}
\description{
Sépare une série en tendance, saisonnalité et résidu.
}
\examples{
x <- ts(rnorm(120) + sin(2*pi*(1:120)/12), frequency = 12)
modele_additif(x)
}
