% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modele_AR.R
\name{modele_auto_regressif}
\alias{modele_auto_regressif}
\title{Modèle auto-régressif AR(p)}
\usage{
modele_auto_regressif(x, p = 1)
}
\arguments{
\item{x}{vecteur numérique ou ts}

\item{p}{ordre du modèle (entier)}
}
\value{
liste avec coefficients et série ajustée
}
\description{
Ajuste un modèle AR simple sur une série temporelle.
}
\examples{
x <- arima.sim(list(ar = 0.6), n = 100)
modele_auto_regressif(x, p = 1)
}
