% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modele_multi.R
\name{modele_multiplicatif}
\alias{modele_multiplicatif}
\title{Décomposition multiplicative}
\usage{
modele_multiplicatif(x)
}
\arguments{
\item{x}{objet ts strictement positif}
}
\value{
objet decompose
}
\description{
Décompose une série en tendance, saisonnalité et résidu (forme multiplicative).
}
\examples{
x <- ts(abs(rnorm(120)) + 10, frequency = 12)
modele_multiplicatif(x)
}
