% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume_serie_temporelle.R
\name{resume_serie}
\alias{resume_serie}
\title{Résumé d’une série temporelle}
\usage{
resume_serie(x)
}
\arguments{
\item{x}{vecteur numérique ou ts}
}
\value{
liste de statistiques
}
\description{
Fournit des statistiques descriptives simples.
}
\examples{
x <- ts(rnorm(50))
resume_serie(x)
}
