% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuillage_quantile.R
\name{seuillage_quantile}
\alias{seuillage_quantile}
\title{Seuillage par quantiles}
\usage{
seuillage_quantile(x, q_low = 0.05, q_high = 0.95)
}
\arguments{
\item{x}{vecteur numérique}

\item{q_low}{quantile bas}

\item{q_high}{quantile haut}
}
\value{
liste avec bornes et indicateur d’extrêmes
}
\description{
Identifie ou corrige les valeurs extrêmes d’une série.
}
\examples{
x <- c(rnorm(100), 10)
seuillage_quantile(x)
}
