% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tendance_lineaire.R
\name{tendance_lineaire}
\alias{tendance_lineaire}
\title{Tendance linéaire}
\usage{
tendance_lineaire(x)
}
\arguments{
\item{x}{vecteur numérique ou ts}
}
\value{
liste avec pente et tendance estimée
}
\description{
Ajuste une droite de régression sur une série temporelle.
}
\details{
La tendance représente l'évolution globale et de long terme d'une série
chronologique. Elle peut être croissante, décroissante ou stable.

Selon les données, la tendance peut être modélisée par une relation linéaire
ou par une relation non linéaire plus flexible.

Identifier la tendance permet de mieux comprendre la dynamique générale
de la série et constitue une étape importante dans la modélisation.
}
\examples{
x <- ts(1:20 + rnorm(20))
tendance_lineaire(x)
}
