% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tendance_non_lineaire.R
\name{tendance_non_lineaire}
\alias{tendance_non_lineaire}
\title{Tendance non linéaire (exponentielle simple)}
\usage{
tendance_non_lineaire(x)
}
\arguments{
\item{x}{vecteur numérique ou ts (valeurs positives conseillées)}
}
\value{
liste avec paramètres et tendance estimée
}
\description{
Ajuste une tendance exponentielle simple sur une série.
}
\examples{
t <- 1:30
x <- 2 * exp(0.05 * t) + rnorm(30)
tendance_non_lineaire(x)
}
