% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_chow.R
\name{test_chow_cassure}
\alias{test_chow_cassure}
\title{Test de Chow (cassure structurelle)}
\usage{
test_chow_cassure(y, x, breakpoint)
}
\arguments{
\item{y}{variable expliquée}

\item{x}{variable explicative}

\item{breakpoint}{position de la cassure}
}
\value{
liste avec statistique F et p-value
}
\description{
Teste un changement de relation linéaire à une date donnée.
}
\examples{
n <- 60
x <- 1:n
y <- c(1 + 0.2*x[1:30], 1 + 0.8*x[31:60]) + rnorm(n)
test_chow_cassure(y, x, breakpoint = 30)
}
