% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_comparaison_moyenne.R
\name{test_comparaison_moyenne}
\alias{test_comparaison_moyenne}
\title{Test de comparaison de moyennes (2 échantillons)}
\usage{
test_comparaison_moyenne(
  x,
  y,
  var.equal = FALSE,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{x, y}{vecteurs numériques}

\item{var.equal}{TRUE si variances supposées égales}

\item{alternative}{"two.sided", "less" ou "greater"}
}
\value{
objet htest (t.test)
}
\description{
Compare les moyennes de deux échantillons indépendants.
}
\details{
Le test de comparaison de moyennes permet d'évaluer si deux échantillons
indépendants ont des moyennes significativement différentes.

Il est couramment utilisé pour comparer deux groupes, par exemple avant/après
une intervention ou entre un groupe témoin et un groupe test.

Selon les hypothèses d'égalité des variances, différentes versions du test
peuvent être appliquées.

L'interprétation repose sur la p-valeur, qui indique si la différence observée
entre les moyennes est statistiquement significative.
}
\examples{
x <- rnorm(20, mean = 5)
y <- rnorm(25, mean = 6)
test_comparaison_moyenne(x, y)
}
