% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_khi_deux.R
\name{test_khi_deux}
\alias{test_khi_deux}
\title{Test du Khi-deux}
\usage{
test_khi_deux(x, p = NULL)
}
\arguments{
\item{x}{table de contingence ou vecteur de comptes}

\item{p}{probabilités théoriques (optionnel)}
}
\value{
objet htest (chisq.test)
}
\description{
Test d’indépendance ou d’adéquation.
}
\details{
Le test du khi-deux est utilisé pour analyser la relation entre deux variables
qualitatives à partir d'un tableau de contingence.

Il permet de tester l'indépendance entre les variables ou l'adéquation
entre des fréquences observées et des fréquences théoriques.

Ce test est largement utilisé dans les études descriptives et exploratoires
impliquant des données catégorielles.
}
\examples{
tab <- matrix(c(20, 10, 15, 25), nrow = 2)
test_khi_deux(tab)
}
