% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_moyenne.R
\name{test_moyenne}
\alias{test_moyenne}
\title{Test de moyenne (1 échantillon)}
\usage{
test_moyenne(x, mu0 = 0, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{x}{vecteur numérique}

\item{mu0}{moyenne sous l’hypothèse nulle H0}

\item{alternative}{"two.sided", "less" ou "greater"}
}
\value{
objet htest (t.test)
}
\description{
Teste si la moyenne d’un échantillon est égale à une valeur donnée.
}
\details{
Le test de moyenne permet de vérifier si la moyenne d'un échantillon est
statistiquement différente d'une valeur de référence donnée.

Ce test est fréquemment utilisé pour comparer une moyenne observée à une
valeur théorique ou attendue.

Selon les hypothèses sur la distribution des données et la taille de l'échantillon,
le test peut être basé sur la loi normale ou la loi de Student.

Le résultat du test repose sur la p-valeur, qui permet de décider si la différence
observée est significative ou non au seuil choisi.
}
\examples{
x <- rnorm(30, mean = 10)
test_moyenne(x, mu0 = 10)
}
