% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_normalite.R
\name{test_normalite}
\alias{test_normalite}
\title{Test de normalité (Shapiro-Wilk)}
\usage{
test_normalite(x)
}
\arguments{
\item{x}{vecteur numérique}
}
\value{
objet htest (shapiro.test)
}
\description{
Vérifie si un échantillon suit une loi normale.
}
\details{
Le test de normalité permet de vérifier si un échantillon peut être considéré
comme issu d'une loi normale.

Cette étape est importante car de nombreux tests statistiques reposent
sur l'hypothèse de normalité des données.

En cas de rejet de l'hypothèse de normalité, il peut être préférable d'utiliser
des méthodes non paramétriques.
}
\examples{
x <- rnorm(30)
test_normalite(x)
}
