% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_proba.R
\name{test_probabilite}
\alias{test_probabilite}
\title{Test de probabilité (binomial exact)}
\usage{
test_probabilite(x, n, p0 = 0.5)
}
\arguments{
\item{x}{nombre de succès}

\item{n}{nombre d'essais}

\item{p0}{probabilité sous H0}
}
\value{
objet htest (binom.test)
}
\description{
Teste si une proportion est égale à une valeur donnée.
}
\examples{
test_probabilite(35, 50, p0 = 0.5)
}
