% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_comparison_variance.R
\name{test_variance}
\alias{test_variance}
\title{Test de comparaison de variances}
\usage{
test_variance(x, y)
}
\arguments{
\item{x, y}{vecteurs numériques}
}
\value{
objet htest (var.test)
}
\description{
Compare les variances de deux échantillons.
}
\examples{
x <- rnorm(20, sd = 2)
y <- rnorm(25, sd = 3)
test_variance(x, y)
}
