% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_variance.R
\name{test_variance_simple}
\alias{test_variance_simple}
\title{Test de variance (1 échantillon)}
\usage{
test_variance_simple(x, sigma0)
}
\arguments{
\item{x}{vecteur numérique}

\item{sigma0}{écart-type sous H0}
}
\value{
liste avec statistique et p-value
}
\description{
Teste si la variance d’un échantillon est égale à une valeur donnée.
}
\details{
Le test de variance permet de vérifier si la dispersion d'un échantillon
est conforme à une valeur de référence ou de comparer les variances de deux
échantillons.

Il est utile pour analyser la stabilité ou la variabilité des données.

Ce test repose généralement sur des hypothèses de normalité des données.
Une variance significativement différente peut indiquer une hétérogénéité
importante dans les observations.
}
\examples{
x <- rnorm(25, sd = 2)
test_variance_simple(x, sigma0 = 2)
}
