% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_d_une_serie.R
\name{visualisation_serie}
\alias{visualisation_serie}
\title{Visualisation simple d’une serie temporelle}
\usage{
visualisation_serie(x)
}
\arguments{
\item{x}{objet ts}
}
\description{
Trace la serie avec un titre automatique.
}
\details{
La visualisation est une étape clé dans l'analyse des séries chronologiques.
Elle permet d'observer l'évolution des données dans le temps et d'identifier
visuellement des phénomènes tels que la tendance, la saisonnalité ou les ruptures.

Une représentation graphique claire facilite l'interprétation des résultats
et guide le choix des méthodes d'analyse à appliquer par la suite.
}
\examples{
x <- ts(rnorm(30), frequency = 12)
visualisation_serie(x)
}
