% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalle_de_conf_mean.R
\name{ic_mean}
\alias{ic_mean}
\title{Intervalle de confiance pour la moyenne}
\usage{
ic_mean(x, conf.level = 0.95)
}
\arguments{
\item{x}{Numeric vector.}

\item{conf.level}{Niveau de confiance (par defaut 0.95).}
}
\value{
Liste contenant :
\describe{
  \item{mean}{moyenne de l'echantillon}
  \item{conf.int}{intervalle de confiance}
  \item{n}{taille de l'echantillon}
}
}
\description{
Calcule un intervalle de confiance pour la moyenne d'un echantillon
en utilisant la loi de Student.
}
\examples{
x <- stats::rnorm(30, mean = 10, sd = 2)
ic_mean(x)
}
