% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalle_de_conf_var.R
\name{ic_variance}
\alias{ic_variance}
\title{Intervalle de confiance pour la variance}
\usage{
ic_variance(x, conf.level = 0.95)
}
\arguments{
\item{x}{Numeric vector.}

\item{conf.level}{Niveau de confiance (par defaut 0.95).}
}
\value{
Liste contenant :
\describe{
  \item{variance}{variance empirique}
  \item{conf.int}{intervalle de confiance}
  \item{n}{taille de l'echantillon}
}
}
\description{
Calcule un intervalle de confiance pour la variance d'un echantillon
en supposant une distribution normale.
}
\examples{
x <- stats::rnorm(25, sd = 2)
ic_variance(x)
}
