% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_automatique_test.R
\name{interpretation_test}
\alias{interpretation_test}
\title{Interprétation automatique d’un test}
\usage{
interpretation_test(p_value, alpha = 0.05)
}
\arguments{
\item{p_value}{p-value du test}

\item{alpha}{niveau de risque (par défaut 5%)}
}
\value{
caractère
}
\description{
Fournit une conclusion simple à partir de la p-value.
}
\examples{
interpretation_test(0.03)
}
