% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/led.R
\name{lissage_exp_double}
\alias{lissage_exp_double}
\title{Lissage exponentiel double (Holt)}
\usage{
lissage_exp_double(x, alpha, beta)
}
\arguments{
\item{x}{vecteur numérique ou ts}

\item{alpha}{lissage du niveau}

\item{beta}{lissage de la tendance}
}
\value{
liste avec niveau, tendance et série ajustée
}
\description{
Modélise le niveau et la tendance d’une série sans saisonnalité.
}
\examples{
x <- c(5, 6, 8, 9, 11)
lissage_exp_double(x, alpha = 0.4, beta = 0.2)
}
