% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les.R
\name{lissage_exp_simple}
\alias{lissage_exp_simple}
\title{Lissage exponentiel simple (SES)}
\usage{
lissage_exp_simple(x, alpha)
}
\arguments{
\item{x}{vecteur numérique ou ts}

\item{alpha}{paramètre de lissage (0 < alpha <= 1)}
}
\value{
liste contenant la série lissée
}
\description{
Lisse une série à l’aide d’un paramètre alpha.
}
\examples{
x <- c(10, 12, 13, 12, 14)
lissage_exp_simple(x, alpha = 0.3)
}
