% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moyenne_mobile.R
\name{moyenne_mobile}
\alias{moyenne_mobile}
\title{Moyenne mobile simple}
\usage{
moyenne_mobile(x, k, centre = TRUE)
}
\arguments{
\item{x}{vecteur numérique ou objet ts}

\item{k}{taille de la fenêtre (entier >= 2)}

\item{centre}{logique, TRUE pour moyenne centrée}
}
\value{
vecteur numérique ou objet ts de même longueur
}
\description{
Moyenne mobile simple
}
\details{
La moyenne mobile est une méthode de lissage des séries chronologiques
consistant à remplacer chaque observation par la moyenne des valeurs
situées dans une fenêtre temporelle de taille fixée.

\strong{Cas impair}

Lorsque l’ordre de la moyenne mobile est impair, on considère une fenêtre
de taille p = 2k + 1, avec k  un entier naturel non nul.
La moyenne mobile centrée d’ordre 2k+1 au temps t est définie par :
\deqn{
M_{2k+1}(t) = \frac{1}{2k+1}\sum_{j=-k}^{k} x_{t+j}
}
pour t = k+1, \ldots, n-k.

Dans ce cas, la fenêtre est symétrique autour de l’instant t ,
ce qui permet un centrage direct de la moyenne mobile et un bon
alignement temporel avec la série initiale.

\strong{Cas pair}

Lorsque l’ordre est pair, on considère une fenêtre de taille p = 2k.
La moyenne mobile ne peut pas être centrée directement car la fenêtre
est située entre deux observations successives.
Le calcul s’effectue alors en deux étapes.

\emph{Étape 1 : calcul d’une moyenne mobile non centrée}
\deqn{
M_{2k}(t) = \frac{1}{2k}\sum_{j=0}^{2k-1} x_{t+j}
}
pour t = 1, \ldots, n-2k+1.

\emph{Étape 2 : centrage de la moyenne mobile}
\deqn{
\widetilde{M}_{2k}(t) = \frac{1}{2}\left(M_{2k}(t) + M_{2k}(t+1)\right)
}

Cette seconde étape permet de recentrer la série lissée dans le temps
afin de garantir un alignement correct avec la série d’origine.
}
\examples{
x <- 1:12
moyenne_mobile(x, k = 3)

ts1 <- ts(rnorm(24), frequency = 12)
moyenne_mobile(ts1, k = 5)
}
