% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moyenne_mobile.R
\name{moyenne_mobile}
\alias{moyenne_mobile}
\title{Moyenne mobile simple}
\usage{
moyenne_mobile(x, k, centre = TRUE)
}
\arguments{
\item{x}{vecteur numérique ou objet ts}

\item{k}{taille de la fenêtre (entier >= 2)}

\item{centre}{logique, TRUE pour moyenne centrée}
}
\value{
vecteur numérique ou objet ts de même longueur
}
\description{
Calcule une moyenne mobile afin de lisser les fluctuations de court terme
d’une série temporelle.
}
\examples{
x <- 1:12
moyenne_mobile(x, k = 3)

ts1 <- ts(rnorm(24), frequency = 12)
moyenne_mobile(ts1, k = 5)
}
