% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tendance_lineaire.R
\name{tendance_lineaire}
\alias{tendance_lineaire}
\title{Tendance linéaire}
\usage{
tendance_lineaire(x)
}
\arguments{
\item{x}{vecteur numérique ou ts}
}
\value{
liste avec pente et tendance estimée
}
\description{
Ajuste une droite de régression sur une série temporelle.
}
\examples{
x <- ts(1:20 + rnorm(20))
tendance_lineaire(x)
}
