% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_khi_deux.R
\name{test_khi_deux}
\alias{test_khi_deux}
\title{Test du Khi-deux}
\usage{
test_khi_deux(x, p = NULL)
}
\arguments{
\item{x}{table de contingence ou vecteur de comptes}

\item{p}{probabilités théoriques (optionnel)}
}
\value{
objet htest (chisq.test)
}
\description{
Test d’indépendance ou d’adéquation.
}
\examples{
tab <- matrix(c(20, 10, 15, 25), nrow = 2)
test_khi_deux(tab)
}
