% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_moyenne.R
\name{test_moyenne}
\alias{test_moyenne}
\title{Test de moyenne (1 échantillon)}
\usage{
test_moyenne(x, mu0 = 0, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{x}{vecteur numérique}

\item{mu0}{moyenne sous l’hypothèse nulle H0}

\item{alternative}{"two.sided", "less" ou "greater"}
}
\value{
objet htest (t.test)
}
\description{
Teste si la moyenne d’un échantillon est égale à une valeur donnée.
}
\examples{
x <- rnorm(30, mean = 10)
test_moyenne(x, mu0 = 10)
}
