% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_normalite.R
\name{test_normalite}
\alias{test_normalite}
\title{Test de normalité (Shapiro-Wilk)}
\usage{
test_normalite(x)
}
\arguments{
\item{x}{vecteur numérique}
}
\value{
objet htest (shapiro.test)
}
\description{
Vérifie si un échantillon suit une loi normale.
}
\examples{
x <- rnorm(30)
test_normalite(x)
}
