% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_variance.R
\name{test_variance_simple}
\alias{test_variance_simple}
\title{Test de variance (1 échantillon)}
\usage{
test_variance_simple(x, sigma0)
}
\arguments{
\item{x}{vecteur numérique}

\item{sigma0}{écart-type sous H0}
}
\value{
liste avec statistique et p-value
}
\description{
Teste si la variance d’un échantillon est égale à une valeur donnée.
}
\examples{
x <- rnorm(25, sd = 2)
test_variance_simple(x, sigma0 = 2)
}
