% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_d_une_serie.R
\name{visualisation_serie}
\alias{visualisation_serie}
\title{Visualisation simple d’une série temporelle}
\usage{
visualisation_serie(x)
}
\arguments{
\item{x}{objet ts}
}
\description{
Trace la série avec un titre automatique.
}
\examples{
x <- ts(rnorm(30), frequency = 12)
visualisation_serie(x)
}
